% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{one_plate}
\alias{one_plate}
\title{The composition of community in a single 24-, 96- or 384-well plate}
\usage{
one_plate(
  plate_type = c("24", "96", "384"),
  plate_name = NULL,
  strain_base = NULL,
  strain_add = NULL,
  strain_sep = "/",
  return_layout = FALSE
)
}
\arguments{
\item{plate_type}{specify plate type, either 24, 96, or 384.}

\item{plate_name}{will be used as the prefix in combination id}

\item{strain_base}{the base strain or community in a single plate}

\item{strain_add}{the id of added strains}

\item{strain_sep}{separator used to join strains into combination}

\item{return_layout}{return the layout of plate, instead of a sample table if TRUE}
}
\value{
a tibble
}
\description{
We use a fast, extensible method to generate multiple SynComs in every
possible composition. The smallest unit is either a 4 x 4 plate (24 well)
for 16 SynComs, a 8x8 plate (96 well) for 64 SynComs, or a 16 x 16
plate (384 well) for 256 SynComs. Based on this,
one can generate as many as plates, and in that cases the combination
of SynComs could be multiplied in batch. For each time we
add an extra strain, the number of total plates will double to fulfill
all additional strain combinations. In other words, the possible
combination of $n$ strain equals to $2^n$.
}
\examples{
  one_plate("24")
  one_plate("96")
  one_plate("384")
}
