% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{sort_combination}
\alias{sort_combination}
\title{Sort combination in a string}
\usage{
sort_combination(strain_combination, strain_sep = "/")
}
\arguments{
\item{strain_combination}{a string or a character vector}

\item{strain_sep}{separator}
}
\value{
a string
}
\description{
String is used as combination name. However, the combination of A and B can be
"A/B" or "B/A", this would make it difficult in comparison of combination names.
By using this function, the name of strains included in a combination will be sorted
alphabetically, so that the combination names can be identical and human-friendly.
}
\examples{
  sort_combination("A1/A11/A2/B3/D5/C11")
  sort_combination(c("A1/A11/A2/B3/D5/C11","A1/A11/A25/B3/D18/C11") )
}
