% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{strain_combination}
\alias{strain_combination}
\title{All the possible combination of multiple strains}
\usage{
strain_combination(strains, strain_sep = "/")
}
\arguments{
\item{strains}{names of strains, can be any integers}

\item{strain_sep}{separator used to join strains into combination}
}
\value{
a character vector
}
\description{
All the possible combination of multiple strains
}
\examples{
  strain_combination(c("A","B","C"))
}
